// src/mail/mail.enum.ts
export enum MailEnum {
  RESET_PASSWORD = 'reset-password',
  BLOCK_EMAIL = 'block-email',
  BLOCK_COMPANY = 'block-company',
  COMMON_INVITATION = 'common-invitation',
  CARRIER_INVITATION = 'carrier-invitation',
  CARRIER_INVITATION_ORDER = 'dispatch-sheet',
  CLIENT_INVITATION = 'client-invitation',
  CLIENT_NEW_LEAD = 'client-new-lead',
  CLIENT_MAIL_CONFIRM = 'client-mail-confirm',
  REQUESTED_CHANGES = 'requested-changes',
}